refinement
when transporting, should not use variant length integer

+++ transport/err.go
@@ -5,11 +5,11 @@ import (
 )
 
 type Error struct {
-	C int
+	C int64
 	M string
 }
 
-func NewErr(code int, err error) *Error {
+func NewErr(code int64, err error) *Error {
 	var msg string
 	if err != nil {
 		msg = err.Error()
@@ -25,6 +25,6 @@ func NoErr() *Error {
 	return &Error{}
 }
 
-func (me *Error) Code() int     { return me.C }
+func (me *Error) Code() int64   { return me.C }
 func (me *Error) Msg() string   { return me.M }
 func (me *Error) Error() string { return fmt.Sprintf("[%d] %v", me.C, me.M) }
