+++ transport/err_test.go
@@ -17,6 +17,6 @@ func TestErrMarshal(t *testing.T) {
 	err = json.Unmarshal(body, &e)
 	ass.Nil(err)
 
-	ass.Equal(0, e.Code())
+	ass.Equal(int64(0), e.Code())
 	ass.Equal("test string", e.Msg())
 }
